+++ R/strat.R
@@ -162,8 +162,7 @@ strat <- function(x,
   txn <- c(NA, diff(posn))
   txn[posn == 1 & is.na(txn)] <- 1
   if (posn[end] == 1) txn[end + 1L] <- -1
-  if (!sum(txn, na.rm = TRUE) == 0) stop("Calculation error - please check validity of data",
-                                         call. = FALSE)
+  if (!sum(txn, na.rm = TRUE) == 0) stop("Calculation error - please check validity of data", call. = FALSE)
 
   logret <- c(diff(log(core[, "open"])), NA)
   if (dir == "short") logret <- -logret